backend done

+++ routes/api.js
@@ -16,11 +16,21 @@ router.get('/notes', function (req, res, next) {
 
 /* POST listing. */
 router.post('/notes/add', (req, res, next) => {
-  console.log(1);
-
+  if (!req.session.user) {
+    return res.send({
+      status: 1,
+      errorMsg: '请先登录'
+    })
+  }
   let note = req.body.note
+  let username = req.session.user.username
+  console.log({
+    text: note,
+    username: username
+  })
   Note.create({
-    text: note
+    text: note,
+    username
   }).then(() => {
     res.send({
       status: 0
@@ -31,37 +41,50 @@ router.post('/notes/add', (req, res, next) => {
       errorMsg: '数据库出错'
     })
   })
-  console.log('add...', note);
-
 });
 
-router.post('/notes/delete', function (req, res, next) {
+router.post('/notes/delete', (req, res, next) => {
+  if (!req.session.user) {
+    return res.send({
+      status: 1,
+      errorMsg: '请先登录'
+    })
+  }
+  let username = req.session.user.username
   Note.destroy({
     where: {
-      id: req.body.id
+      id: req.body.id,
+      username: username
     }
   }).then(() => {
     res.send({
       status: 0
     })
   }).catch(() => {
-    sres.send({
+    res.send({
       status: 1,
-      error: '数据库错误'
+      errorMsg: '数据库错误'
     })
   })
 });
 
 router.post('/notes/edit', function (req, res, next) {
+  if (!req.session.user) {
+    return res.send({
+      status: 1,
+      errorMsg: '请先登录'
+    })
+  }
   let body = req.body
+  let username = req.session.user.username
   Note.update({
     text: body.note
   }, {
     where: {
-      id: body.id
+      id: body.id,
+      username: username
     }
   }).then(() => {
-    console.log(arguments)
     res.send({
       status: 0
     })
